Version 1.0.4

+++ mypyc_aot/mypyc_aot.py
@@ -1,5 +1,5 @@
-import sys, os, io, string, base64, random
-import pickle, sysconfig, tempfile
+import sys, os, io, string, base64, random, shutil
+import pickle, sysconfig, tempfile, threading
 from hashlib import sha256
 import subprocess
 try:
@@ -10,9 +10,10 @@ from setuptools import setup
 from mypy.defaults import CACHE_DIR as _mypy_cache_basename
 from mypyc.build import mypycify
 
-__all__ = ["mypyc_aot_nocache", "mypyc_aot"]
-_home = os.getenv("USERPROFILE") if sys.platform == "win32" \
-        else os.getenv("HOME")
+__all__ = ["mypyc_aot_nocache", "mypyc_aot", "clear_all_cache",
+           "clear_intermediate_cache"]
+_home = os.environ["USERPROFILE"] if sys.platform == "win32" \
+        else os.environ["HOME"]
 CACHE_DIR = os.path.join(_home, ".mypyc_aot_cache")
 try:
     __import__("zstandard")
@@ -25,6 +26,19 @@ except ImportError:
         DEFAULT_COMP_METHOD = None
 
 _config = {}
+
+class RedirectedStream:
+    def __init__(self, orig_stream, stream):
+        self.orig_stream = orig_stream
+        self.stream = stream
+        self.thread_id = threading.get_ident()
+    def write(self, data):
+        if threading.get_ident() == self.thread_id: # 是当前线程
+            return self.stream.write(data)
+        else:
+            return self.orig_stream.write(data)
+    def __getattr__(self, attr):
+        return getattr(self.orig_stream, attr) # 返回orig_stream的属性和方法
 
 class NoCompression:
     @staticmethod
@@ -56,9 +70,11 @@ def captured_popen(stream):
             return super().__init__(*args, **kw)
         def wait(self, *args, **kw):
             result = super().wait(*args, **kw)
+            if self.stdout is None or self.stderr is None:
+                return result
             if not self.stdout.closed and not self.stderr.closed:
                 # _communicate: 避免递归
-                stdout, stderr = self._communicate(None, None, None)
+                stdout, stderr = self._communicate(None, None, None) # type: ignore
                 if stdout is not None:
                     stream.write(stdout.decode("utf-8", "replace"))
                 if stderr is not None:
@@ -69,15 +85,17 @@ def captured_popen(stream):
 def mypyc_aot_nocache(pycode: str, prefix="mypyc_aot", cache_dir = None,
                       quiet=True, compiler: str | None = None,
                       opt_level="3", strict_dunder_typing=True,
-                      experimental_features=False):
+                      experimental_features=False, debug_level="0",
+                      **kw):
     if cache_dir is None:
         temp_dir = tempfile.gettempdir()
     else:
         temp_dir = os.path.join(cache_dir, "cache")
-    py_path = os.path.join(temp_dir, f"{rand_name(16)}-mypyc-aot")
+    random_name = rand_name(16)
+    py_path = os.path.join(temp_dir, f"{random_name}-mypyc-aot")
     os.makedirs(py_path, exist_ok=True)
 
-    mod_name = f"{prefix}_{rand_name(16)}"
+    mod_name = f"{prefix}_{random_name}"
     pyfile = os.path.join(py_path, f"{mod_name}.py")
     with open(pyfile, "w", encoding="utf-8") as f:
         f.write(pycode)
@@ -106,15 +124,19 @@ def mypyc_aot_nocache(pycode: str, prefix="mypyc_aot", cache_dir = None,
         sys.argv[1:] = ["build_ext", "--inplace"]
         if quiet:
             sys.argv.append("-q")
-        sys.stderr = sys.stdout = stream
-        sys.path.append(py_path)
+        sys.stdout = RedirectedStream(sys.stdout, stream)
+        sys.stderr = RedirectedStream(sys.stderr, stream)
         os.chdir(py_path)
+        new_sys_path = [*sys.path, py_path]
+        sys.path = new_sys_path
         try:
-            setup(name = mod_name,
+            setup(
+                name = mod_name,
                 ext_modules = mypycify([pyfile, "--check-untyped-defs"],
                                        opt_level=opt_level,
                                        strict_dunder_typing=strict_dunder_typing,
-                                       experimental_features=experimental_features),
+                                       experimental_features=experimental_features,
+                                       debug_level=debug_level, **kw),
             )
         except SystemExit as err:
             sys.stdout, sys.stderr = prev_stdout, prev_stderr
@@ -123,6 +145,7 @@ def mypyc_aot_nocache(pycode: str, prefix="mypyc_aot", cache_dir = None,
                     f"Nonzero exit ({err.code}), outputs:\n{stream.getvalue()}")\
                     from None
 
+        sys.path = new_sys_path
         module =  __import__(mod_name)
         module.__mypyc_aot_path__ = get_module_path(mod_name)
         return module
@@ -160,11 +183,13 @@ reversible compress() and decompress() function:
         with open(cache_file, "rb") as f:
             data = pickle.loads(mod.decompress(f.read()))
         py_path = os.path.join(cache_dir, "cache",
-                               f"{hash}{compression_method}-mypyc")
+                               f"{hash}{compression_method}-lib")
         os.makedirs(py_path, exist_ok=True)
         filename = os.path.join(py_path, data["name"])
-        with open(filename, "wb") as f:
-            f.write(data["data"])
+        if not os.path.isfile(filename):
+            with open(filename, "wb") as f:
+                f.write(data["data"])
+
         prev_path = sys.path.copy()
         try:
             sys.path.append(py_path)
@@ -185,3 +210,14 @@ reversible compress() and decompress() function:
                 })
             ))
         return compiled
+
+def clear_all_cache():
+    if os.path.isdir(CACHE_DIR):
+        shutil.rmtree(CACHE_DIR)
+
+def clear_intermediate_cache():
+    # 仅清理中间产物，不清理缓存
+    if os.path.isdir(os.path.join(CACHE_DIR, "cache")):
+        shutil.rmtree(os.path.join(CACHE_DIR, "cache"))
+    if os.path.isdir(os.path.join(CACHE_DIR, ".mypy_cache")):
+        shutil.rmtree(os.path.join(CACHE_DIR, ".mypy_cache"))
